Fixed Dog.cs entity, added DbSet context Dogs, moved RegisterDog model, implemented RegisterDog in UserService.cs
