+++ dotnet_walkADog_backend/Helpers/DataContext.cs
@@ -20,4 +20,6 @@ public class DataContext : DbContext
 
     public DbSet<User> Users { get; set; }
     public DbSet<TrainerData> TrainerData { get; set; }
+
+    public DbSet<Dog> Dogs { get; set; }
 }
