+++ dotnet_walkADog_backend/Models/Dogs/RegisterDog.cs
